requirements.txt
biopython==1.83
numpy==1.24.4
